fix: movement within border underflow of unsigned integer

+++ image-viewer/include/direction.h
@@ -10,6 +10,6 @@ typedef enum {
   RIGHT,
 } direction_t;
 
-void direction_move_xy(direction_t direction, uint16_t *x, uint16_t *y, int16_t amount);
+void direction_move_xy(direction_t direction, int32_t *x, int32_t *y, int16_t amount);
 
 #endif // __DIRECTION_H__
